lace-platform: Add hwid interface module
Pull `find_edid`, `find_smbios_tables`, the DTB find/install helpers
and the CHID-based platform-compatibility logic out of the crate root
and platform `mod.rs` files into a dedicated `hwid` module.

The top-level `hwid.rs` owns the portable `platform_compatible_*` /
`chid_sources` helpers (moved from `lib.rs`) and re-exports the
platform-specific `find_dtb` / `find_edid` / `find_smbios_tables` /
`install_dtb`. Per-platform: new `bios/hwid.rs` (SMBIOS scan from the
legacy memory range), `efi/dtb.rs` becomes `efi/hwid.rs` and folds in
the EDID + SMBIOS finders that used to live in `efi/mod.rs`, new
`mock/hwid.rs` placeholder.

Take the opportunity to make `efi::open_protocol_exclusive` private and
rename it to `open_first_protocol_exclusive` to disambiguate it from
`uefi::boot::open_protocol_exclusive`, which now has a real call site
in `efi/hwid.rs`. Rename `efi::mem::efi_mem_init` to a plain `init` for
consistency with the other platforms.

Update `lace-stubble` import paths from `lace_platform::dtb::*` and
`lace_platform::platform_compatible_*` to `lace_platform::hwid::*`.
